Moved ChatMessage object
